refactor: [FE] setPixel() 위치 수정
- Intent
useCanvas() 내 state에 의존하지 않은 함수 utils로 옮김
