+++ frontend/src/utils/canvas.ts
@@ -29,3 +29,11 @@ export const getPixelColor = (imageData: any, x: number, y: number) => {
 export function isSameColor(a: Uint8ClampedArray, b: Uint8ClampedArray) {
     return a[0] === b[0] && a[1] === b[1] && a[2] === b[2] && a[3] === b[3];
 }
+
+export const setPixel = (imageData: any, x: number, y: number, color: Uint8ClampedArray) => {
+    const offset = getPixelOffset(imageData, x, y);
+    imageData.data[offset + 0] = color[0];
+    imageData.data[offset + 1] = color[1];
+    imageData.data[offset + 2] = color[2];
+    imageData.data[offset + 3] = color[0];
+};
